+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ==
 
+## [1.1.0] - 2022-04-19
+### Added
+- New entering and leaving state responses added to StateListenerBase.
+- Previous state now tracked and sent with state change events.
+
 ## [1.0.0] - 2022-01-14
 ### Added
 - Initial creation of Unity-State.